week2/smart-data-analyst/backend/requirements.txt
fastapi
uvicorn[standard]
sse-starlette
langchain
langchain-community
langgraph
langchain-qwq
sqlalchemy
aiosqlite
python-dotenv
pydantic
python-multipart
